style: fix formatting for CI

+++ src/Task1Factory.sol
@@ -16,12 +16,12 @@ contract Task1Factory is ReentrancyGuard {
         WBNB = IWETH(_wbnb);
     }
 
-    function buyExactOut(
-        uint256 amountOut,
-        address[] calldata path,
-        address to,
-        uint256 deadline
-    ) external payable nonReentrant returns (uint256[] memory amounts) {
+    function buyExactOut(uint256 amountOut, address[] calldata path, address to, uint256 deadline)
+        external
+        payable
+        nonReentrant
+        returns (uint256[] memory amounts)
+    {
         require(path.length == 2, "Only direct pairs supported");
         require(path[0] == address(WBNB), "path[0]!=WBNB");
         require(block.timestamp <= deadline, "DEADLINE");
@@ -29,28 +29,23 @@ contract Task1Factory is ReentrancyGuard {
         address pair = IPancakeFactory(factory).getPair(path[0], path[1]);
         require(pair != address(0), "PAIR_NOT_EXISTS");
 
-        (uint112 reserve0, uint112 reserve1, ) = IPancakePair(pair)
-            .getReserves();
+        (uint112 reserve0, uint112 reserve1,) = IPancakePair(pair).getReserves();
 
         address token0 = IPancakePair(pair).token0();
 
-        (uint reserveIn, uint reserveOut) = path[0] == token0
-            ? (reserve0, reserve1)
-            : (reserve1, reserve0);
+        (uint256 reserveIn, uint256 reserveOut) = path[0] == token0 ? (reserve0, reserve1) : (reserve1, reserve0);
 
         require(amountOut < reserveOut, "INSUFFICIENT_LIQUIDITY");
-        uint numerator = reserveIn * amountOut * 10000;
-        uint denominator = (reserveOut - amountOut) * 9975;
-        uint amountIn = (numerator / denominator) + 1;
+        uint256 numerator = reserveIn * amountOut * 10000;
+        uint256 denominator = (reserveOut - amountOut) * 9975;
+        uint256 amountIn = (numerator / denominator) + 1;
 
         require(amountIn <= msg.value, "EXCESSIVE_INPUT");
 
         WBNB.deposit{value: amountIn}();
         WBNB.transfer(pair, amountIn);
 
-        (uint amount0Out, uint amount1Out) = path[0] == token0
-            ? (uint(0), amountOut)
-            : (amountOut, uint(0));
+        (uint256 amount0Out, uint256 amount1Out) = path[0] == token0 ? (uint256(0), amountOut) : (amountOut, uint256(0));
 
         IPancakePair(pair).swap(amount0Out, amount1Out, to, new bytes(0));
