+++ src/Task2Liquidity.sol
@@ -17,35 +17,28 @@ contract Task2Liquidity is ReentrancyGuard {
         router = _router;
         WBNB = _router.WETH();
     }
-    function buyAndPool(
-        uint256 amountTokenDesired,
-        address[] calldata path,
-        uint256 deadline
-    ) external payable nonReentrant returns (uint256 liquidity) {
+
+    function buyAndPool(uint256 amountTokenDesired, address[] calldata path, uint256 deadline)
+        external
+        payable
+        nonReentrant
+        returns (uint256 liquidity)
+    {
         require(path.length == 2, "BAD_PATH");
         require(path[0] == WBNB, "path[0]!=WBNB");
 
-        uint[] memory amounts = router.swapETHForExactTokens{value: msg.value}(
-            amountTokenDesired,
-            path,
-            address(this),
-            deadline
-        );
+        uint256[] memory amounts =
+            router.swapETHForExactTokens{value: msg.value}(amountTokenDesired, path, address(this), deadline);
         uint256 bnbSpent = amounts[0];
         uint256 bnbLeft = msg.value - bnbSpent;
 
         address pair = IPancakeFactory(router.factory()).getPair(path[1], WBNB);
         require(pair != address(0), "NO_PAIR");
-        (uint112 reserve0, uint112 reserve1, ) = IPancakePair(pair)
-            .getReserves();
-        (uint112 reserveToken, uint112 reserveBNB) = (path[1] < WBNB)
-            ? (reserve0, reserve1)
-            : (reserve1, reserve0);
+        (uint112 reserve0, uint112 reserve1,) = IPancakePair(pair).getReserves();
+        (uint112 reserveToken, uint112 reserveBNB) = (path[1] < WBNB) ? (reserve0, reserve1) : (reserve1, reserve0);
 
         uint256 optimalTokenAmount = (bnbLeft * reserveToken) / reserveBNB;
-        uint256 usedTokenAmount = optimalTokenAmount > amountTokenDesired
-            ? amountTokenDesired
-            : optimalTokenAmount;
+        uint256 usedTokenAmount = optimalTokenAmount > amountTokenDesired ? amountTokenDesired : optimalTokenAmount;
         uint256 usedBNBAmount = (usedTokenAmount * reserveBNB) / reserveToken;
 
         uint256 minTokenAmount = (usedTokenAmount * 995) / 1000;
@@ -54,13 +47,8 @@ contract Task2Liquidity is ReentrancyGuard {
         IERC20(path[1]).approve(address(router), usedTokenAmount);
 
         // Add liquidity
-        (, , liquidity) = router.addLiquidityETH{value: usedBNBAmount}(
-            path[1],
-            usedTokenAmount,
-            minTokenAmount,
-            minBNBAmount,
-            msg.sender,
-            deadline
+        (,, liquidity) = router.addLiquidityETH{value: usedBNBAmount}(
+            path[1], usedTokenAmount, minTokenAmount, minBNBAmount, msg.sender, deadline
         );
 
         _refundDust(IERC20(path[1]), msg.sender);
